Refactor `Vitals` to use `float` for temperature, add `simulate` logic to `Krzys` for dynamic vitals updates, and schedule periodic simulation.

+++ src/main/java/micgrz/krzysbackend/vitals/Vitals.java
@@ -15,6 +15,6 @@ public class Vitals {
 
     private int energia;
     private int tetno;
-    private int temperatura;
+    private float temperatura;
     private Nastroj nastroj;
 }
